% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAcceso.R
\docType{data}
\name{LastUpdate}
\alias{LastUpdate}
\title{Fecha de la última actualización de meta-datos del paquete}
\format{An object of class \code{POSIXct} (inherits from \code{POSIXt}) of length 1.}
\source{
\url{http://infra.datos.gob.ar/catalog/modernizacion/dataset/1/distribution/1.2/download/series-tiempo-metadatos.csv}
}
\usage{
LastUpdate
}
\description{
(De los datos Listado)
}
\keyword{datasets}
