% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAcceso.R
\name{List}
\alias{List}
\title{List
Buscar series con el pattern deseado}
\usage{
List(PATTERN = "*")
}
\arguments{
\item{PATTERN}{Pattern de búsqueda en la descripción de la serie}
}
\value{
Tibble con las series disponibles que con descripción coincidente
}
\description{
List
Buscar series con el pattern deseado
}
\examples{
Listado <- List("PIB")
}
