% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAcceso.R
\docType{data}
\name{Listado}
\alias{Listado}
\title{Listado de Series del Portal de Hacienda}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14341 rows and 24 columns.}
\source{
\url{http://infra.datos.gob.ar/catalog/modernizacion/dataset/1/distribution/1.2/download/series-tiempo-metadatos.csv}
}
\usage{
Listado
}
\description{
Listado completo de series.
uso: data(Listado)
}
\keyword{datasets}
