% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListadoMake.R
\name{ListadoMake}
\alias{ListadoMake}
\title{Armar la base de metadatos de series
Buscar series con el pattern deseado - descarga la serie}
\usage{
ListadoMake()
}
\arguments{
\item{PATTERN}{Pattern de búsqueda en la descripción de la serie}
}
\value{
Tibble con las series disponibles que con descripción coincidente

Listado <- List("PIB")
}
\description{
Armar la base de metadatos de series
Buscar series con el pattern deseado - descarga la serie
}
