% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAcceso.R
\docType{package}
\name{PortalHacienda}
\alias{PortalHacienda}
\alias{PortalHacienda-package}
\title{PortalHacienda: Interface R a la API de datos del Ministerio de Hacienda}
\description{
PortalHacienda: Interface R a la API de datos del Ministerio de Hacienda
}
\section{PortalHacienda functions}{

\code{Search} busca las series en la base de meta-datos incluida (descargada del portal oficial)

\code{Search_online} busca las series descargando la última versión del paquete de meta-datos (10mb aprox)

\code{Get} obtiene las series desde la API

\code{Forecast} extiende las series obtenidas con un modelo auto-detectado por el paquete **forecast**
}

