% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAcceso.R
\name{Search}
\alias{Search}
\title{Buscar series en el último archivo de meta-datos descargado
Busca series con el pattern deseado - descarga la serie}
\usage{
Search(PATTERN = "*")
}
\arguments{
\item{PATTERN}{Pattern de búsqueda en la descripción de la serie}
}
\value{
Tibble con las series disponibles que con descripción coincidente
}
\description{
Buscar series en el último archivo de meta-datos descargado
Busca series con el pattern deseado - descarga la serie
}
\examples{
Listado <- SeriesSearch("PIB")
}
