% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAcceso.R
\name{Forecast}
\alias{Forecast}
\title{Forecast
Extender series con forecast de auto.arima}
\usage{
Forecast(SERIE, N = 6)
}
\arguments{
\item{SERIE}{XTS a extender}

\item{N}{Cantidad de períodos a extender}
}
\value{
XTS con la serie expandida e intervalos de confianza al 95%
}
\description{
Forecast
Extender series con forecast de auto.arima
}
\examples{
X <- PortalHacienda::Forecast(Get("138.1_PAPDE_0_M_41"),12)
}
