% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAcceso.R
\name{Get}
\alias{Get}
\title{Acceder a la API del Portal de Datos}
\usage{
Get(series, start_date = NULL, end_date = NULL)
}
\arguments{
\item{series}{ID de la serie a obtener}

\item{start_date}{Fecha de inicio}

\item{end_date}{Fecha de final}
}
\value{
Un objeto XTS con la serie seleccionada en ID
}
\description{
\code{Get} devuelve la serie seleccionada en ID
}
\examples{
X <- Get("138.1_PAPDE_0_M_41")
}
