% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAcceso.R
\name{Search}
\alias{Search}
\title{Buscar series en el archivo de meta-datos descargado con el paquete
Si se desea utilizar la versión online más actualizada usar Search_online (sólo con conexión a internet)}
\usage{
Search(PATTERN = "*")
}
\arguments{
\item{PATTERN}{Pattern de búsqueda en la descripción de la serie}
}
\value{
Tibble con las series disponibles que con descripción coincidente
}
\description{
Buscar series en el archivo de meta-datos descargado con el paquete
Si se desea utilizar la versión online más actualizada usar Search_online (sólo con conexión a internet)
}
\examples{
Listado <- Search("Tipo de Cambio")
}
